requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.20
pydantic==2.5.0
slowapi==0.1.9
aiofiles==23.2.1
jinja2==3.1.2
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.2
requests==2.31.0